fix build add pom and main class
