--- src/main/java/com/springbloog/blog/controller/PostResource.java
@@ -1,55 +1,0 @@
-package com.springbloog.blog.controller;
-
-import com.springbloog.blog.entity.Post;
-import com.springbloog.blog.payload.PostDto;
-import com.springbloog.blog.payload.PostResponse;
-import com.springbloog.blog.service.PostService;
-import com.springbloog.blog.utils.AppConstants;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/posts")
-public class PostResource {
-    //constructor based dependency injection
-    private PostService postService;  //here we are making loose coupling by adding interface (not tight coupling via adding class)
-    @Autowired    //if class is spring bean and have only   constructor then can use @Autowire
-    public PostResource(PostService postService) {
-        this.postService = postService;
-    }
-
-    @PostMapping
-    private ResponseEntity<PostDto> createPost(@RequestBody PostDto postDto){
-        return new ResponseEntity<>(postService.createPost(postDto), HttpStatus.CREATED);
-    }
-
-    @GetMapping
-    private ResponseEntity<PostResponse> getAllPosts(
-            @RequestParam(value = "pageNo",defaultValue = AppConstants.DEFAULT_PAGE_NUMBER,required = false) int pageNo,
-            @RequestParam(value = "pageSize",defaultValue = AppConstants.DEFAULT_PAGE_SIZE,required = false) int pageSize,
-            @RequestParam(value= "sortBy",defaultValue = AppConstants.DEFAULT_SORT_BY,required = false) String sortBy,
-            @RequestParam(value = "sortDir",defaultValue = AppConstants.DEFAULT_SORT_DIRECTION,required = false) String sortDir
-    ){
-        return ResponseEntity.ok(postService.getAllPosts(pageNo,pageSize,sortBy,sortDir));
-    }
-
-    @GetMapping("/{id}")
-    private ResponseEntity<PostDto> getPostById(@PathVariable("id") Long id){
-        return ResponseEntity.ok(postService.getPostById(id));
-    }
-    @PutMapping("/{id}")
-    private ResponseEntity<PostDto> updatePostId(@RequestBody PostDto postDto,@PathVariable(name = "id") Long id){
-        return ResponseEntity.ok(postService.updatePost(postDto,id));
-    }
-    @DeleteMapping("/{id}")
-    private ResponseEntity<String> deletePostById(@PathVariable(name="id") Long id){
-        postService.deletePostById(id);
-        return new ResponseEntity<>("Post deleted successfully",HttpStatus.OK);
-    }
-
-}
