--- src/main/java/com/springbloog/blog/entity/Post.java
@@ -1,27 +1,0 @@
-package com.springbloog.blog.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(schema = "myBlog",name = "Posts",
-uniqueConstraints = {@UniqueConstraint(columnNames = "title")})
-public class Post {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(name = "title", nullable = false)
-    private String title;
-    @Column(name = "description",nullable = false)
-    private String description;
-    @Column(name="content",nullable = false)
-    private String content;
-
-}
