--- src/main/java/com/springbloog/blog/payload/PostDto.java
@@ -1,13 +1,0 @@
-package com.springbloog.blog.payload;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class PostDto {
-    private Long id;
-    private String title;
-    private String description;
-    private String content;
-}
