+++ src/main/java/com/springboot/blog/SpringbootBlogRestApiApplication.java
@@ -1,0 +1,13 @@
+package com.springboot.blog;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringbootBlogRestApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringbootBlogRestApiApplication.class, args);
+	}
+
+}
