+++ src/main/java/com/springboot/blog/exception/ResourceNotFoundException.java
@@ -1,0 +1,30 @@
+package com.springboot.blog.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+//Response status annotation cause springboot to respond with the specific Http Status code whenever this
+//exception thrown from controller
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException{
+    private String resourceName;
+    private String fieldName;
+    private Long fieldValue;
+
+    public ResourceNotFoundException(String resourceName,String fieldName,Long fieldValue) {
+        super(String.format("%s not found with %s : %s ",resourceName,fieldName,fieldValue));
+        this.resourceName = resourceName;
+        this.fieldName = fieldName;
+        this.fieldValue = fieldValue;
+    }
+
+    public String getResourceName() {
+        return resourceName;
+    }
+    public String getFieldName() {
+        return fieldName;
+    }
+    public Long getFieldValue() {
+        return fieldValue;
+    }
+
+}
