+++ src/main/java/com/springboot/blog/payload/PostDto.java
@@ -1,0 +1,13 @@
+package com.springboot.blog.payload;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class PostDto {
+    private Long id;
+    private String title;
+    private String description;
+    private String content;
+}
