+++ src/main/java/com/springboot/blog/payload/PostResponse.java
@@ -1,0 +1,18 @@
+package com.springboot.blog.payload;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class PostResponse {
+    private List<PostDto> content;
+    private int pageNo;
+    private int pageSize;
+    private long totalElements;
+    private int totalPages;
+    private boolean last;
+}
