+++ src/main/java/com/springboot/blog/service/PostService.java
@@ -1,0 +1,13 @@
+package com.springboot.blog.service;
+
+import com.springboot.blog.payload.PostDto;
+import com.springboot.blog.payload.PostResponse;
+
+public interface PostService {
+    PostDto createPost(PostDto postDto);
+    PostResponse getAllPosts(int pageNo, int pageSize,String sortBy,String sortDir);
+    PostDto getPostById(Long id);
+    PostDto updatePost(PostDto postDto,Long id);
+    void deletePostById(Long id);
+
+}
