+++ src/main/java/com/springboot/blog/utils/AppConstants.java
@@ -1,0 +1,10 @@
+package com.springboot.blog.utils;
+
+import lombok.Builder;
+
+public class AppConstants {
+    public static final String DEFAULT_PAGE_NUMBER="O";
+    public static final String DEFAULT_PAGE_SIZE="10";
+    public static final String DEFAULT_SORT_BY="id";
+    public static final String DEFAULT_SORT_DIRECTION="asc";
+}
