+++ src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.datasource.url = jdbc:mysql://localhost:3306/myblog?useSSL=false&serverTimezone=UTC
+spring.datasource.username = root
+spring.datasource.password =
+
+# hibernate properties
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQLDialect
+
+# Hibernate ddl auto (create, create-drop, validate, update)
+spring.jpa.hibernate.ddl-auto = update
